requirements_uv.txt
anyio==4.9.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
attrs==25.3.0
autopep8==2.3.2
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.3.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
executing==0.8.3
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
identify==2.6.9
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-server==2.15.0
jupyter-server-mathjax==0.2.6
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mistune==3.1.3
mpmath==1.3.0
mypy==1.15.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nbqa==1.9.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
platformdirs==3.10.0
pre-commit==4.2.0
prometheus-client==0.21.1
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.13.0
pycparser==2.22
pygments==2.15.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
ruff==0.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
stack-data==0.2.0
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenize-rt==6.1.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
virtualenv==20.30.0
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
